Complete #1600 by moving User-Defined Type tests out of RestApiv2Test as well
